refactor(chat): extract background task send helpers
